fix:修复发送朋友圈返回的id解析为float64

+++ api/khan/v1/sns.go
@@ -202,8 +202,8 @@ type SnsSendTextRequest struct {
 }
 
 type SnsSendTextResponse struct {
-	Id         int64  `json:"id"`
-	UserName   string `json:"userName"`
-	NickName   string `json:"nickName"`
-	CreateTime int64  `json:"createTime"`
+	Id         float64 `json:"id"`
+	UserName   string  `json:"userName"`
+	NickName   string  `json:"nickName"`
+	CreateTime int64   `json:"createTime"`
 }
